requirements.txt
asgiref==3.8.1
boto3==1.34.148
botocore==1.34.148
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
